encapsulated logic into it's own class
